requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.6
django-extensions==4.1
django-import-export==4.3.7
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.1
whitenoise==6.9.0

#pip freeze > requirements.txt
#python manage.py collectstatic

